Add CI and dependabot configuration, fix broken builds
Besides a few minor clippy lints, the documentation for this crate was
completely stale and not compiling.  By testing this in CI we ensure
users will be able to run the code that we provide to them.

Note that strangely most functions in the "builder pattern"
`Intent` are suffixed `with_` (because they _move_ `Intent`) except
`set_class_name()` and `into_chooser(_with_title)()`.  Is this intended
or an oversight?